requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.7
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
